CCC SSP-1424: UP-4665: Fix layout.json HTTP 500 issue when tenant admin deletes portlet that's in one of their layouts.
Conflicts:
	uportal-war/src/main/java/org/jasig/portal/layout/dlm/RDBMDistributedLayoutStore.java
